duzo dodalem, ekrany wyboru klasy, mapy, victory screen

+++ ProjektPSIO/ParallaxLayer.cpp
@@ -63,10 +63,10 @@ void ParallaxLayer::draw(
     float viewHeight =
         camera.getSize().y;
 
-    float textureWidth =
+    int textureWidth =
         texture.getSize().x;
 
-    float textureHeight =
+    int textureHeight =
         texture.getSize().y;
 
     float scaleX =
